Correcção erro em função add na matriz de doubles - Recebia um int em vez de double. Copy/paste fail.

+++ src/lei/li3/g50/utilitarios/Matriz_Double_12x2.java
@@ -145,7 +145,7 @@ public class Matriz_Double_12x2 {
         }
     }
 
-    public void addValorMesTipoCompra(Mes mes, TipoCompra tipo_compra, int valor) {
+    public void addValorMesTipoCompra(Mes mes, TipoCompra tipo_compra, double valor) {
         if (tipo_compra != TipoCompra.AMBOS) {
             this.matriz[mes.getIndiceArray()][tipo_compra.getIndiceArray()] += valor;
             this.total += valor;
